+++ superset/assets/src/components/TableSelector.jsx
@@ -215,10 +215,11 @@ export default class TableSelector extends React.PureComponent {
     return this.renderSelectRow(
       <AsyncSelect
         dataEndpoint={
-          '/databaseasync/api/' +
-          'read?_flt_0_expose_in_sqllab=1&' +
-          '_oc_DatabaseAsync=database_name&' +
-          '_od_DatabaseAsync=asc'
+          '/api/v1/database/?q=' +
+          '(keys:!(none),' +
+          'columns:!(id,database_name,backend),' +
+          'filters:!((col:expose_in_sqllab,opr:eq,value:!t)),' +
+          'order_columns:database_name,order_direction:asc)'
         }
         onChange={this.onDatabaseChange}
         onAsyncError={() => this.props.handleError(t('Error while fetching database list'))}
